The bug is dead. The origin was mfSplitIdempotent; at the end, after splitting the idempotent, we have to suspend everything once. This was being done everywhere except for the last step (producing non-square matrices). Correcting this leads to the correct results in all orders for the trefoil.
There are many additional checks in blow, linkhom and mfweb added during the process of tracking down this bug. Some can be removed soon, but I suggest leaving most of them; there is now a very comprehensive system of checks in place should any further typos or errors creep in. All checks can be toggled on or off in blowFlags (turning them off speeds things up, but not a great deal).