Probable fix of problem with 6_1, 6_3 and L7a7. In the earlier fix, compilation was adjusted to skip pairs of vertices with no edges between them, but this is a bad idea: it means we go "off script" from the compilation strategy. But then when compiling morphisms, we still follow the original compilation strategy which means that we are using the wrong matrix factorisations.
The correct fix is to not allow compilation strategies where this occurs, i.e. where one tries to fuse vertices with no edges between them. Such comp strats are now called "invalid" and there is compilationStratVerify in mfweb.lib which checks this.

Now linkhom.lib also tries to use only valid compilation strategies. It begins with the naive one, just taking the vertices in the order given by the encoding, but if this is invalid it tries to permute the vertices in order to find a valid strategy.

The default strategies for 6_1, 6_3 and L7a7 are invalid, but the code now quickly finds a valid permutation and starts compiling webs. I believe at the end there should now be no problem with the maps, but I haven't checked this.